Use sync.RWMutex instead of normal sync.Mutex
Using an `RLock` on read operation prevents mutex to block the current goroutine during _READONLY_ durations while it should block the process during a _WRITE_ operation, `sync.RWMutex` would be the best solution. As an HTTP server may get to see a huge load of requests, it could save us a lot of time if we look up at the bigger frame counting multiple instantaneous requests. Also, an approach with the least mutex lock time would be the most optimum, whenever possible, and hence I propose individual `read/write` locks for the `(*RateLimiter) Limit(string)bool` method with respect to the stated scenario.